# Conflicts:
#	src/main/resources/static/js/agregarPacientes.js

+++ src/main/java/com/example/Arevalo_Saibene_Nosicoski/repository/IDomicilioRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IDomicilioRepository  extends JpaRepository <Domicilio,Integer> {
+public interface IDomicilioRepository  extends JpaRepository <Domicilio,Long> {
 
 }
